chore(clerk-react): Move errorThrower.ts to /errors path
