+++ packages/react/src/components/withSession.tsx
@@ -2,8 +2,8 @@ import type { SessionResource } from '@clerk/types';
 import React from 'react';
 
 import { useSessionContext } from '../contexts/SessionContext';
+import { errorThrower } from '../errors/errorThrower';
 import { hocChildrenNotAFunctionError } from '../errors/messages';
-import { errorThrower } from '../utils';
 
 export const withSession = <P extends { session: SessionResource }>(
   Component: React.ComponentType<P>,
