+++ packages/react/src/components/withUser.tsx
@@ -2,8 +2,8 @@ import type { UserResource } from '@clerk/types';
 import React from 'react';
 
 import { useUserContext } from '../contexts/UserContext';
+import { errorThrower } from '../errors/errorThrower';
 import { hocChildrenNotAFunctionError } from '../errors/messages';
-import { errorThrower } from '../utils';
 
 export const withUser = <P extends { user: UserResource }>(Component: React.ComponentType<P>, displayName?: string) => {
   displayName = displayName || Component.displayName || Component.name || 'Component';
