+++ packages/react/src/hooks/useAssertWrappedByClerkProvider.ts
@@ -1,6 +1,6 @@
 import { useAssertWrappedByClerkProvider as useSharedAssertWrappedByClerkProvider } from '@clerk/shared/react';
 
-import { errorThrower } from '../utils';
+import { errorThrower } from '../errors/errorThrower';
 
 export const useAssertWrappedByClerkProvider = (source: string): void => {
   useSharedAssertWrappedByClerkProvider(() => {
