+++ packages/react/src/isomorphicClerk.ts
@@ -36,6 +36,7 @@ import type {
   Without,
 } from '@clerk/types';
 
+import { errorThrower } from './errors/errorThrower';
 import { unsupportedNonBrowserDomainOrProxyUrlFunction } from './errors/messages';
 import type {
   BrowserClerk,
@@ -45,7 +46,7 @@ import type {
   HeadlessBrowserClerkConstructor,
   IsomorphicClerkOptions,
 } from './types';
-import { errorThrower, isConstructor, loadClerkJsScript } from './utils';
+import { isConstructor, loadClerkJsScript } from './utils';
 
 const SDK_METADATA = { name: PACKAGE_NAME, version: PACKAGE_VERSION };
