+++ packages/react/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './childrenUtils';
-export * from './errorThrower';
 export * from './isConstructor';
 export { loadClerkJsScript } from './loadClerkJsScript';
 export * from './useMaxAllowedInstancesGuard';
